fix: Resolve Netlify deployment build failure
This commit fixes the Netlify build process, which was failing due to issues with serverless function bundling and module compatibility.

- **Isolate Development Dependencies:** The `log` function has been moved from `server/vite.ts` to its own file (`server/log.ts`). This prevents development-only dependencies like Vite from being included in the serverless function bundle, which was a primary cause of the build failure.
- **Configure Netlify Function Bundler:** The `netlify.toml` file has been updated to explicitly set the `node_bundler` to `esbuild` and externalize the `express` module. This ensures that the serverless function is bundled as an ES Module (ESM), resolving conflicts with `import.meta` and top-level `await` syntax.
- **Add Missing Dependency:** The `@babel/preset-typescript` package has been added as a `devDependency` to resolve a "Could not resolve" error that was occurring during the build process.
