+++ netlify.toml
@@ -3,6 +3,12 @@
   publish = "dist/public"
   functions = "netlify/functions"
 
+[functions]
+  node_bundler = "esbuild"
+
+[functions."api"]
+  external_node_modules = ["express"]
+
 [[redirects]]
   from = "/api/*"
   to = "/.netlify/functions/api/:splat"
